refactor: 각 컴포넌트 단일 책임원책으로 재구조
